pyproject.toml
[project]
name = "pixe"
version = "0.5.5"
description = "A digital helper to keep your files neat and tidy"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache License, Version 2.0"}
keywords = ["archive", "photos", "organize"]
authors = [
    {name = "Chris Wells", email = "chris@ithuna.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Topic :: System :: Archiving",
]
dependencies = [
    "Click>=8.1.3,<8.2",
    "Pillow>=9.4.0,<9.5",
    "piexif>=1.1.3,<1.2",
]

[project.optional-dependencies]
tests = [
    "pytest>=7.2.1,<7.3",
    "pytest-cov>=4.0.0,<4.1",
    "flake8>=6.0.0,<6.1",
]

[project.urls]
homepage = "https://github.com/ithuna/pixe.git"

[project.scripts]
pixe = "pixe:cli"